Added the 'evaluate' concept
I've moved the hand tests into new test file HandTest, and added a test
for 'evaluate'.

This 'evaluate' thing is something that I think will be a big help to
us, but it isn't easy to implement so feel free to do the minimum you
can.  You can also change the test if it happens to look a bit weird or
is missing any necessary .toCards.